Module Desugaring__Count_generics_instantiations

This module is used to count the number of different concrete type parameters generic classes are instantiated with.

val count_generics_instantiations_program : Typing.Typed_ast.program -> (Ast.Ast_types.Class_name.t * Ast.Ast_types.type_expr list) list