Module Desugaring__Desugar_program

This module is responsible for desugared the typed AST. The desugared AST simplifies the expressions in the typed ASTs It provides the invariant that there is no variable shadowing in the desugared AST

val desugar_program : Typing.Typed_ast.program -> Desugaring.Desugared_ast.program Core.Or_error.t

val pprint_desugared_ast : Stdlib.Format.formatter -> Desugaring.Desugared_ast.program -> unit

Given a formatter and desugared AST, pretty-print the AST - useful for debugging