Module Desugaring__Desugared_ast

The desugared AST consists of the typed AST but simplified.

for loops have been desugared into while loops.

Shadowing is not permitted, so variables have been given unique monikers

Block expressions have just been replaced by an expr list, since no shadowing so block scope not necessary

type identifier =

| Variable of Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t

| ObjField of Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.Field_name.t
first type is of the object, second is of field

type expr =

| Integer of Ast.Ast_types.loc * int
no need for type_expr annotation as obviously TEInt

| Boolean of Ast.Ast_types.loc * bool
no need for type_expr annotation as obviously TEBool

| Identifier of Ast.Ast_types.loc * identifier
Type information associated with identifier

| Constructor of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Class_name.t * constructor_arg list

| Let of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t * expr

| Assign of Ast.Ast_types.loc * Ast.Ast_types.type_expr * identifier * expr

| Consume of Ast.Ast_types.loc * identifier
Type is associated with the identifier

| MethodApp of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Var_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.Method_name.t * expr list

| FunctionApp of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.Function_name.t * expr list

| Printf of Ast.Ast_types.loc * string * expr list
no need for type_expr annotation as obviously TEVoid

| FinishAsync of Ast.Ast_types.loc * Ast.Ast_types.type_expr * async_expr list * expr list
overall type is that of the expr on the current thread - since forked exprs' values are ignored

| If of Ast.Ast_types.loc * Ast.Ast_types.type_expr * expr * expr list * expr list
If ___ then ___ else ___ - type is that of the branch exprs

| While of Ast.Ast_types.loc * expr * expr list
While ___ do ___ ; - no need for type_expr annotation as type of a loop is TEVoid

| BinOp of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.bin_op * expr * expr

| UnOp of Ast.Ast_types.loc * Ast.Ast_types.type_expr * Ast.Ast_types.un_op * expr

and constructor_arg =

| ConstructorArg of Ast.Ast_types.type_expr * Ast.Ast_types.Field_name.t * expr

Constructor arg consists of a field and the expression being assigned to it (annotated with the type of the expression)

and async_expr =

| AsyncExpr of Ast.Ast_types.Var_name.t list * expr list

Async exprs have a precomputed list of their free variables (passed as arguments when they are spawned as thread)

type function_defn =

| TFunction of Ast.Ast_types.Function_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.param list * expr list

Function defn consists of the function name, return type, the list of params, and the body expr block of the function

type method_defn =

| TMethod of Ast.Ast_types.Method_name.t * Ast.Ast_types.type_expr * Ast.Ast_types.param list * Ast.Ast_types.Region_name.t list * expr list

Method defn consists the method name, return type, the list of params, the region affected and the body expr block of the function

type class_defn =

| TClass of Ast.Ast_types.Class_name.t * Ast.Ast_types.region list * Ast.Ast_types.field_defn list * method_defn list

Class definitions consist of the class name, its capability regions and the fields and methods in the class

type program =

| Prog of class_defn list * function_defn list * expr list

Each bolt program defines the classes,followed by functions, followed by the main expression block to execute.