Module Desugaring__Free_obj_vars_expr

val free_obj_vars_expr : Typing.Typed_ast.class_defn list -> Typing.Typed_ast.expr -> (Ast.Ast_types.Var_name.t * Ast.Ast_types.Class_name.t * Ast.Ast_types.capability list) list

Return a list of the object free variables in an expr and their associated classes and capabilities

val free_obj_vars_block_expr : Typing.Typed_ast.class_defn list -> Typing.Typed_ast.block_expr -> (Ast.Ast_types.Var_name.t * Ast.Ast_types.Class_name.t * Ast.Ast_types.capability list) list

Return a list of the object free variables in a block expr and their associated classes and capabilities